Убрал создание репозитория в отдельный файл, спрятал константы
