+++ app/src/main/java/com/geekbrains/Constans.kt
@@ -1,0 +1,7 @@
+package com.geekbrains
+
+
+const val TOTAL_COUNT_EXTRA = "TOTAL_COUNT_EXTRA"
+const val BASE_URL = "https://api.github.com"
+const val FAKE = "FAKE"
+
